[#157] Introduced "generate-annotations" Maven phase
